--- a/hw10_program_optimization/bench.txt
+++ b/hw10_program_optimization/bench.txt
@@ -1,0 +1,6 @@
+=== RUN   TestGetDomainStat_Time_And_Memory
+    stats_optimization_test.go:46: time used: 198.602176ms / 300ms
+    stats_optimization_test.go:47: memory used: 10Mb / 30Mb
+--- PASS: TestGetDomainStat_Time_And_Memory (3.27s)
+PASS
+ok  	github.com/filatkinen/hw-test/hw10_program_optimization	3.271s